shoppinglyxapi/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
dj-database-url==0.5.0

Django==3.2
djangorestframework==3.13.1
django-admin-interface==0.19.0
django-colorfield==0.6.3
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
fonttools==4.30.0
gunicorn==20.1.0

# heroku==0.1.3

idna==3.3
kiwisolver==1.3.2

matplotlib==3.4

numpy==1.22.3
packaging==21.3
Pillow==9.0.1
psycopg2-binary==2.9.3
pyparsing==3.0.7

python-dateutil==2.7
pytz==2022.1
django-on-heroku==1.1.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.9
waitress==2.1.1
whitenoise==6.0.0
importlib-metadata==4.11.3
markdown==3.3.6
zipp==3.7.0
django-filter==21.1
#django-cors-headers-3.11.0